Segmented local-crawler into multiple classes
