+++ Makefile
@@ -1,4 +1,4 @@
-.PHONY: clean start stop help
+.PHONY: clean start stop help start-debug test
 
 stop:
 	docker-compose down
@@ -19,14 +19,16 @@ start-debug:
 
 clean:
 	rm -rf ./crawl_results
-	docker system prune --all
+	docker system prune
 
 test:
 	DEBUG_MODE=true TEST_CRAWL=true docker-compose up --build -d
 
 help:
 	@echo "Available commands:"
-	@echo "  make start        - Start services in normal mode"
-	@echo "  make start-debug  - Start services in debug mode"
-	@echo "  make stop         - Stop services"
-	@echo "  make clean        - Clean up Docker resources"
+	@echo "  make start        - Starts crawler on all 8 batches of sites with debug mode turned off"
+	@echo "  make start-debug  - Starts crawler on all 8 batches of sites with debug mode turned on"
+	@echo "  make test  	   - Starts crawler on test batch of sites with debug mode turned on"
+	@echo "  make stop         - Stops crawler"
+	@echo "  make clean        - Removes crawl_results and cleans up docker artifacts related to crawler"
+
