+++ selenium-optmeowt-crawler/constants.js
@@ -1,0 +1,8 @@
+module.exports = {
+    UNRECOVERABLE_ERRORS: ['InsecureCertificateError', 'WebDriverError', 'HumanCheckError'],
+    PAGE_LOAD_TIMEOUT: 35000,
+    SCRIPT_TIMEOUT: 30000,
+    RETRY_DELAY: 3000,
+    API_BASE_URL: 'http://rest_api:8080',
+    CRALWER_BROWSER_URL: 'http://crawl_browser:4444'
+};
