+++ selenium-optmeowt-crawler/index.js
@@ -1,0 +1,10 @@
+const WebCrawler = require('./WebCrawler')
+
+// Start the crawler
+const crawler = new WebCrawler(process.argv.slice(2));
+
+(async () => {
+  await crawler.start();
+})();
+  
+
