+++ selenium-optmeowt-crawler/run_crawl.sh
@@ -7,9 +7,9 @@ touch ./crawl_results/"$TIMESTAMP"/error-logging/error-logging.json
 
 
 if [ "$TEST_CRAWL" = "true" ]; then
-  node local-crawler.js $TIMESTAMP 1 -1
+  node index.js $TIMESTAMP 1 -1
 else
-  node local-crawler.js $TIMESTAMP 0 $CRAWL_ID
+  node index.js $TIMESTAMP 0 $CRAWL_ID
 fi
 
 curl -o ./crawl_results/"$TIMESTAMP"/analysis.json "http://rest_api:8080/analysis"
